refactor(storage): Split schema mod into multiple sub-mods
